requirements.txt
tqdm
Pillow
tensorflow
scipy
numpy
matplotlib